Restructured to make installable via pip install
moved package into its own directory
moved examples into its own directory
added setup.py and pyproject.toml
modified import statements to match new structure
