+++ README.md
@@ -5,11 +5,9 @@ hanabi\_learning\_environment is a research platform for Hanabi experiments. The
 ### Getting started
 ```
 sudo apt-get install g++         # if you don't already have a CXX compiler
-sudo apt-get install cmake       # if you don't already have CMake
 sudo apt-get install python-pip  # if you don't already have pip
-pip install cffi                 # if you don't already have cffi
-cmake .
-make
+pip install .                    # or pip install git+repo_url to install directly from github
+
 python rl_env_example.py         # Runs RL episodes
 python game_example.py           # Plays a game using the lower level interface
 ```
