+++ examples/game_example.cc
