+++ examples/rl_env_example.py
@@ -17,9 +17,9 @@ from __future__ import print_function
 
 import sys
 import getopt
-import rl_env
-from agents.random_agent import RandomAgent
-from agents.simple_agent import SimpleAgent
+from hanabi_learning_environment import rl_env
+from hanabi_learning_environment.agents.random_agent import RandomAgent
+from hanabi_learning_environment.agents.simple_agent import SimpleAgent
 
 AGENT_CLASSES = {'SimpleAgent': SimpleAgent, 'RandomAgent': RandomAgent}
