+++ hanabi_learning_environment/agents/simple_agent.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 """Simple Agent."""
 
-from rl_env import Agent
+from hanabi_learning_environment.rl_env import Agent
 
 
 class SimpleAgent(Agent):
