+++ hanabi_learning_environment/hanabi_lib/canonical_encoders.cc
