+++ hanabi_learning_environment/hanabi_lib/hanabi_history_item.h
