+++ hanabi_learning_environment/hanabi_lib/hanabi_observation.cc
