+++ hanabi_learning_environment/hanabi_lib/observation_encoder.h
