+++ hanabi_learning_environment/hanabi_lib/util.cc
